fix: leverage ActiveModel to signal errors on widgets
This means the controller simply instantiates widget and has them
checked (`widget.validate!`) on render. This removes a lot of
unfortunate code in the controller and the headache that the flash had
become.

There is probably a better way to extend ActiveModel: I'm not
sure why `.valid?` had to be explicitly rewritten + using
`validates_with` can probably does the error adding for us.